feat(autoComplete): Add showOnEmpty option
Add an option to allow the tag list to show when input is empty.
This feature is useful when the user wants to see an unfiltered
list of all available suggestions before typing.

The Source expression must be able to handle a null query
if this option is used. Partially addresses #54