feat(tagsInput): Add support for array of objects
Drop support for array of strings and add support for array of objects as
a list of tags. Introduce a new option, display-property, so the user can
set what property should be used as the tag text. The option defaults to
"text".

BREAKING CHANGE: From now on, array of strings are no longer supported. In
order to keep some backward compatibility a one-time conversion from an
array of strings into an array of objects is available.

Closes #46.