feat(autoComplete): down arrow hotkey option to open suggestion box if not shown
Add the down arrow downkey event as a hotkey that opens the suggestion box if the
suggestion box is not visible. This feature is important because it gives the user a method
for browsing the list of tags without needing to filter.

Requires the source function to return a list when input is null. Addresses #54