feat(tagsInput): Add visual feedback for invalid tags
Add visual feedback for invalid tags so one knows why a tag cannot be
added. Make the directive invalid when it loses focus and there is some
leftover text that couldn't be turned into a tag (either because it's
invalid or the addOnBlur option is off). That behavior can be turned off
by setting the newly created allowLeftoverText option to true.

Closes #77.