tell $timeout to use $apply in suggestionList.load() to keep view in sync
When autocomplete minLength is set to 1, typing 1 character does not result in a the suggestion being rendered. You can also try typing multiple characters really fast (or set debounce to a value like 300 or 500) and see that the first batch of quickly typed character yields no suggestion list either. The debouncing mechanism causes suggestion list updates to miss their turns the $digest cycle. Setting the 3rd parameter to $timeout to true will ensure $apply is called so that the view can render the suggestion list in a timely manner.