feat(tagsInput): Add ng-disabled support
Add an option for the user to disable the tags input component. Set ng-disabled to true will prevent any changes to the tags defined within the model. The field can not be focused any more. Anchors to remove tags will be disabled as well as the "add new" placeholder input.

Closes #102.