feat(tagsInput): add newTagText option
Add an option for binding to the text in the input element.
This feature is important because it allows developers to know when
the input has changed as well the state of the input at the time of
the change. An example use case is doing custom validations that
regular expressions couldn't satisfy.

Closes #197.