feat(tagsInput): Add support for adding extra attributes to the .tags html element
Ads extra attributes to the .tags element in the template, useful to add extra directives or any other HTML tag needed to that element. The value should be an object where the keys are the attribute and the value the attribute value. For example tags-extra-attrs="{myelement: 'myvalue'}". Defaults to null.