feat(tagsInput): Add experimental support for array of strings
Add a new option called useStrings that enables experimental support
for array of strings in addiiton to array of objects. This change only
affects the way the model bound to the directive is manipulated.
Internally an array of objects is still used to store existing tags,
and events related to tag operations still expose tags as objects.